refactor(hangar): make the ledger append and its publish one operation
The site-8 fix was an append followed by a publish: two statements a
later edit separates, which is this exact defect for the third time.
append_and_publish takes what the convergence call site has and does
both, so they cannot come apart without changing the function.

The guarantee is now stated in terms the next person can check. Counting
write sites is what made seven successive counts wrong, because that set
is unbounded and grows with the code. The classifier renders exactly
nine event_type tokens, so a row can only diverge live-from-durable if
something minted one of those nine, and every minter of the nine is
behind acp_transcript: the reducer's seven through the sink's chunk
door, the lifecycle markers through its marker door, the one direct
appender through this function.

One exception is named rather than fixed: acp.transcript_truncated is
minted inside the store writer's buffer cap, below the layer that knows
a stream exists. It is unfixable by this design and is the same
buffer-overflow window the live-before-durable ordering already
qualifies the equality with from the other side.

acp_pool no longer names FleetProviderEventRepo at all.
