+++ ainb-tui/crates/ainb-hangar-daemon/src/acp_transcript.rs
@@ -7,11 +7,25 @@
 //! per-MODULE, so a sign on the door inside `acp_pool.rs` is a sign, and the
 //! same omission had already walked past it twice.
 //!
-//! **That is the exact scope of the guarantee, and it is narrower than "every
-//! durable transcript row".** `FleetProviderEventRepo::append` is public and a
-//! caller can reach the ledger without touching a `StoreWriter` at all —
-//! `converge_dirty_session` does, and no wrap here can stop it. Such a caller
-//! must publish through [`publish_repo_row`], and that one IS a convention.
+//! # The guarantee, in terms you can check in one grep
+//!
+//! `AcpClassifier::classify_value` renders exactly NINE `event_type` tokens and
+//! drops everything else, so a row can only diverge live-from-durable if
+//! something minted one of those nine. **Every minter of those nine is behind
+//! this module**: the reducer's seven reach the ledger only through
+//! [`TranscriptSink::chunk`], the lifecycle markers only through
+//! [`TranscriptSink::lifecycle`], and the one caller that appends straight to
+//! the ledger goes through [`append_and_publish`].
+//!
+//! Counting WRITE SITES is what made seven successive counts wrong: that set is
+//! unbounded and grows with the code. Counting TOKEN MINTERS bounds it, because
+//! the token set is nine strings in one `match`.
+//!
+//! One exception, and it is unfixable by design rather than an oversight:
+//! `acp.transcript_truncated` is minted INSIDE `StoreWriter::enforce_buffer_cap`,
+//! below the layer that knows a stream exists, so it reaches durable without
+//! streaming. That is the same buffer-overflow window the live-before-durable
+//! ordering already qualifies the equality with, from the other side.
 
 use std::time::Duration;
 
@@ -19,8 +33,12 @@ use ainb_acp::reducer::TranscriptChunk;
 use ainb_acp::store_writer::{HighWater, Lifecycle, StoreWriter};
 use ainb_hangar_proto::events::MessageKind;
 use ainb_hangar_proto::transcript::AcpClassifier;
-use ainb_hangar_store::repo::fleet_provider_event::FleetProviderEventError;
+use ainb_hangar_store::repo::fleet_provider_event::{
+    FleetProviderEventError, FleetProviderEventRepo, FleetProviderEventRow, NewFleetProviderEvent,
+};
+use sqlx::SqlitePool;
 
+use crate::events::EventSink;
 use crate::runner::RunStream;
 
 /// One session's transcript, live and durable, behind one door.
@@ -50,23 +68,58 @@ pub(crate) struct TranscriptSink {
     pub(crate) tool_calls: u32,
 }
 
-/// Publish ONE row that was written straight to the repo, past the sink.
+/// Bind a session's live task stream, or `None` when the scope is not a task's.
 ///
-/// `converge_dirty_session` mints its `turn_interrupted` marker with
-/// `FleetProviderEventRepo::append` because it is a free function with no actor
-/// and no writer, so the sink's privacy cannot reach it — a durable row that no
-/// wrap can catch. It still has to stream, and it is the one that matters most:
-/// convergence is the operator-stop and adapter-death path, so without this the
-/// live pane ends without the interruption and re-opening the ticket shows a
-/// line that was never streamed.
+/// The scope convention is the whole discriminator: `acp_task` mints
+/// `task:<id>` and nothing else may (`fleet/acp_session_create` refuses the
+/// prefix), so `strip_prefix` is an exact test rather than a heuristic.
+///
+/// Owned arguments so a caller holding a `&mut` actor can await it: the actor
+/// is `Send` but not `Sync`, so a future borrowing it is not spawnable.
+pub(crate) async fn bind_task_stream(
+    pool: SqlitePool,
+    events: EventSink,
+    scope_key: String,
+) -> Option<RunStream> {
+    let task_id = scope_key.strip_prefix(crate::acp_task::TASK_SCOPE_PREFIX)?;
+    let workspace_id = crate::acp_task::workspace_for_scope(&pool, &scope_key).await?;
+    RunStream::bind(&events, &workspace_id, task_id)
+}
+
+/// Append one row STRAIGHT to the ledger and publish it live, in that order,
+/// as one operation.
+///
+/// `converge_dirty_session` mints its `turn_interrupted` marker here because it
+/// is a free function with no actor and no writer, so [`TranscriptSink`]'s
+/// privacy cannot reach it: a renderable durable row that no wrap around
+/// `StoreWriter` can catch. Convergence is the operator-stop and adapter-death
+/// path, so without the publish the live pane ends without the interruption and
+/// re-opening the ticket shows a line nobody saw arrive.
+///
+/// Both halves are ONE function, not two statements at the call site, for the
+/// reason this defect has now appeared three times: two statements are
+/// separable by a later edit, and every previous instance was written by
+/// someone doing something legitimate next to one of them.
 ///
 /// A fresh classifier is correct here and not a shortcut: a lifecycle marker
 /// carries everything it renders, unlike a `tool_call_update` that needs the
 /// call before it.
-pub(crate) fn publish_repo_row(stream: &RunStream, event_type: &str, raw_payload: &str) {
-    for (kind, body) in AcpClassifier::default().classify_row(event_type, raw_payload) {
-        stream.line(kind, body);
+pub(crate) async fn append_and_publish(
+    pool: &SqlitePool,
+    events: &EventSink,
+    scope_key: &str,
+    row: &NewFleetProviderEvent,
+) -> Result<FleetProviderEventRow, FleetProviderEventError> {
+    let stored = FleetProviderEventRepo::append(pool, row).await?;
+    if let Some(stream) =
+        bind_task_stream(pool.clone(), events.clone(), scope_key.to_string()).await
+    {
+        for (kind, body) in AcpClassifier::default().classify_row(&row.event_type, &row.raw_payload)
+        {
+            stream.line(kind, body);
+        }
     }
+    Ok(stored)
 }
 
 impl TranscriptSink {
